+++ src/ui_control.h
@@ -10,6 +10,7 @@
 extern "C" {
 #endif
 
+#include "common.h"
 #include "UIMode.h"
 
 typedef enum _button_status_t {
@@ -30,11 +31,13 @@ void ui_clear_btn_evt();
 
 UIMode* getUIMode(ui_mode_enm_t ui_mode_enm);
 
-void ui_init();
+void ui_init(board_type_t board_type);
 ui_mode_enm_t ui_update();
 ui_mode_enm_t ui_force_update(ui_mode_enm_t ui_mode_enm);
 uint16_t ui_get_idle_count();
 uint32_t ui_set_center_switch_for_wakeup(bool flg);
+
+board_type_t ui_get_board_type();
 
 #ifdef __cplusplus
 }